+++ include/WorldObjects/BezierCurve.hpp
@@ -14,11 +14,11 @@ public:
 
     void drawCurve() const;
 
-    Vec eval_arc(double arc) const;
-    Vec eval_t(double t) const;
+    Vec eval_arc(float arc) const;
+    Vec eval_t(float t) const;
 
-    Vec eval_deriv_arc(double arc) const;
-    Vec eval_deriv_t(double t) const;
+    Vec eval_deriv_arc(float arc) const;
+    Vec eval_deriv_t(float t) const;
 
     void loadFile(const std::string &filename);
     void evalMaxMin();
@@ -26,10 +26,10 @@ public:
     bool drawCage = false;
     bool drawPath = false;
 
-    double getXmin() const { return m_xMin; }
-    double getXmax() const { return m_xMax; }
-    double getZmin() const { return m_zMin; }
-    double getZmax() const { return m_zMax; }
+    float getXmin() const { return m_xMin; }
+    float getXmax() const { return m_xMax; }
+    float getZmin() const { return m_zMin; }
+    float getZmax() const { return m_zMax; }
 
     void setVec(std::vector<Vec> v) { m_points = v; }
     std::vector<Vec> getVec() const { return m_points; }
@@ -42,17 +42,17 @@ protected:
 private:
     std::vector<Vec> m_points;
     // Table to cache curve points.
-    mutable std::vector<double> m_cache_t;
-    mutable std::vector<double> m_cache_arc;
+    mutable std::vector<float> m_cache_t;
+    mutable std::vector<float> m_cache_arc;
     mutable std::vector<Vec> m_cache_pos;
     mutable std::vector<Vec> m_cache_deriv;
 
-    std::pair<double, size_t> findInCache(double t) const;
-    Vec evalCubic(Vec p0, Vec p1, Vec p2, Vec p3, double t) const;
-    Vec evalCubicDeriv(Vec p0, Vec p1, Vec p2, Vec p3, double t) const;
+    std::pair<float, size_t> findInCache(float t) const;
+    Vec evalCubic(Vec p0, Vec p1, Vec p2, Vec p3, float t) const;
+    Vec evalCubicDeriv(Vec p0, Vec p1, Vec p2, Vec p3, float t) const;
 
-    double m_xMin;
-    double m_xMax;
-    double m_zMin;
-    double m_zMax;
+    float m_xMin;
+    float m_xMax;
+    float m_zMin;
+    float m_zMax;
 };
