+++ include/WorldObjects/Lighting/Spotlight.hpp
@@ -8,15 +8,15 @@ public:
     Spotlight() = default;
     virtual ~Spotlight() = default;
 
-    double exponent() const { return m_spot_exp; }
-    void exponent(double exp) { m_spot_exp = exp; }
+    float exponent() const { return m_spot_exp; }
+    void exponent(float exp) { m_spot_exp = exp; }
 
-    double cutoff() const { return m_spot_cutoff; }
-    void cutoff(double angle) { m_spot_cutoff = angle; }
+    float cutoff() const { return m_spot_cutoff; }
+    void cutoff(float angle) { m_spot_cutoff = angle; }
 
 protected:
-    double m_spot_exp    = 1.0;
-    double m_spot_cutoff = 45.0;
+    float m_spot_exp    = 1.0;
+    float m_spot_cutoff = 45.0;
 
-    virtual void update(double t, double dt) override;
+    virtual void update(float t, float dt) override;
 };
